fixed race condition causing 'null' being used as fasthash
if the update process was interrupted between listing all files and calculating the hashes, we would get 'null' strings as fasthash, and because they remain in the index they never get updated, and would go on to infect the hoard.

fixed by removing non-existent keys from the filesystem index toml.